requirements.txt
pyaudio 
wave
azure-cognitiveservices-speech 
python-dotenv